Broker for SPI message to moisture and motor done. Test to be done.

+++ firmware/WaterCtrl_v1/cubemx/Inc/moistureMeasure.h
@@ -16,15 +16,11 @@
 #ifndef MOISTUREMEASURE_H
 #define MOISTUREMEASURE_H
 
-#include <stm32f1xx_hal.h>
-
-#include <log.h>
-
 int initMoistureMeasure(TIM_HandleTypeDef*);
 
 int deinitMoistureMeasure(void);
 
-uint_fast64_t getSensorFrequency(int sensor);
+uint32_t getSensorFrequency(int sensor);
 
 void printMoisture();
